move reply routine to preprocessor
